@@ -36,7 +36,7 @@
   },
   "dependencies": {
     "@types/dojo": "1.9.41",
-    "sass": "^1.17.0",
-    "typescript": "3.3"
+    "node-sass": "4.14.1",
+    "typescript": "4.2.4"
   }
 }
