@@ -14,18 +14,13 @@
  * limitations under the License.
  *
  */
-/// <amd-dependency path="esri/core/tsSupport/declareExtendsHelper" name="__extends" />
-/// <amd-dependency path="esri/core/tsSupport/decorateHelper" name="__decorate" />
-
-import { subclass, declared, property } from "esri/core/accessorSupport/decorators";
-import { tsx, renderable } from "esri/widgets/support/widget";
-
-import Widget = require("esri/widgets/Widget");
+import { subclass, property } from "esri/core/accessorSupport/decorators";
+import { tsx } from "esri/widgets/support/widget";
+import Widget from "esri/widgets/Widget";
 
 @subclass("widgets/Toggle")
-class Toggle extends declared(Widget) {
+class Toggle extends Widget {
   @property()
-  @renderable()
   active: boolean = false;
 
   render() {
